Move introspection, prometheus setup from agent top level package to a new addons package;
